REFACTOR : Data type of column name of us census csv data

+++ src/main/java/com/bridgelabz/statecensusanalyser/CSVUSCensus.java
@@ -5,13 +5,16 @@ import com.opencsv.bean.CsvBindByName;
 public class CSVUSCensus {
     @CsvBindByName(column = " State Id", required = true)
     public String stateId;
+
     @CsvBindByName(column = "State", required = true)
     public String usState;
+
     @CsvBindByName(column = "Population", required = true)
-    public int usPopulation;
+    public long usPopulation;
+
     @CsvBindByName(column = "Total area", required = true)
     public double totalArea;
+
     @CsvBindByName(column = "Population Density", required = true)
     public double populationDensity;
-
 }
